Create an executable script for Quaternary Symmetry detection
